Move container types to Container namespace
