+++ src/Service/Lexicon/V1/ComponentGenerator/Field/ArrayComponentGenerator.php
@@ -7,7 +7,7 @@ use Blugen\Service\Lexicon\ArraySerialization\ArraySerializationContext;
 use Blugen\Service\Lexicon\ArraySerialization\ArraySerializationContributor;
 use Blugen\Service\Lexicon\GeneratorInterface;
 use Blugen\Service\Lexicon\V1\Property;
-use Blugen\Service\Lexicon\V1\Schema\Field\ArraySchema;
+use Blugen\Service\Lexicon\V1\Schema\Container\ArraySchema;
 use Nette\PhpGenerator\ClassType;
 
 class ArrayComponentGenerator implements GeneratorInterface, ArraySerializationContributor
