+++ src/Service/Lexicon/V1/Schema/Container/ParamsSchema.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Blugen\Service\Lexicon\V1\Schema\Field;
+namespace Blugen\Service\Lexicon\V1\Schema\Container;
 
 use Blugen\Service\Lexicon\SchemaInterface;
 use Blugen\Service\Lexicon\V1\Property;
@@ -8,12 +8,14 @@ use Blugen\Service\Lexicon\V1\Schema;
 use Blugen\Service\Lexicon\V1\Traits\ArrayableTrait;
 use Blugen\Service\Lexicon\V1\Traits\RawSchemaAccessorTrait;
 
-class ObjectSchema implements SchemaInterface
+class ParamsSchema implements SchemaInterface
 {
     use ArrayableTrait;
     use RawSchemaAccessorTrait;
 
-    public function __construct(private readonly SchemaInterface $schema)
+    public function __construct(
+        private readonly SchemaInterface $schema,
+    )
     {
     }
 
@@ -24,7 +26,7 @@ class ObjectSchema implements SchemaInterface
 
     public function description(): ?string
     {
-        return $this->schema->description() ?? null;
+        return $this->schema->description();
     }
 
     public function __get(string $name): mixed
@@ -32,27 +34,27 @@ class ObjectSchema implements SchemaInterface
         return $this->schema->__get($name);
     }
 
-    public function properties(): array
-    {
-        $nullable = $this->nullable() ?? [];
-        $required = $this->required() ?? [];
-        $properties = $this->__get('properties') ?? [];
-
-        return array_map(fn (string $name, array $rawSchema) => new Property(
-            $name,
-            new Schema($rawSchema),
-            in_array($name, $nullable, true),
-            in_array($name, $required, true),
-        ), array_keys($properties), $properties);
-    }
-
+    /**
+     * @return string[]|null
+     */
     public function required(): ?array
     {
         return $this->__get('required');
     }
 
-    public function nullable(): ?array
+    /**
+     * @return Property[]
+     */
+    public function properties(): array
     {
-        return $this->__get('nullable');
+        $required = $this->required() ?? [];
+        $properties = $this->__get('properties');
+
+        return array_map(fn (string $name, array $rawSchema) => new Property(
+            $name,
+            new Schema($rawSchema),
+            ! in_array($name, $required, true),
+            in_array($name, $required, true),
+        ), array_keys($properties), $properties);
     }
 }
