+++ src/Service/Lexicon/V1/TypeSpecificDefinition/Primary/RecordTypeDefinition.php
@@ -5,7 +5,7 @@ namespace Blugen\Service\Lexicon\V1\TypeSpecificDefinition\Primary;
 use Blugen\Service\Lexicon\DefinitionInterface;
 use Blugen\Service\Lexicon\LexiconInterface;
 use Blugen\Service\Lexicon\V1\Schema;
-use Blugen\Service\Lexicon\V1\Schema\Field\ObjectSchema;
+use Blugen\Service\Lexicon\V1\Schema\Container\ObjectSchema;
 use Blugen\Service\Lexicon\V1\Traits\ArrayableTrait;
 use Blugen\Service\Lexicon\V1\Traits\DefinitionTrait as DefinitionTrait;
