+++ src/Service/Syntax/Factory/SchemaFactory.php
@@ -11,8 +11,8 @@ use Blugen\Service\Lexicon\V1\Schema\Concrete\CidLinkSchema;
 use Blugen\Service\Lexicon\V1\Schema\Concrete\IntegerSchema;
 use Blugen\Service\Lexicon\V1\Schema\Concrete\NullSchema;
 use Blugen\Service\Lexicon\V1\Schema\Concrete\StringSchema;
-use Blugen\Service\Lexicon\V1\Schema\Field\ArraySchema;
-use Blugen\Service\Lexicon\V1\Schema\Field\ObjectSchema;
+use Blugen\Service\Lexicon\V1\Schema\Container\ArraySchema;
+use Blugen\Service\Lexicon\V1\Schema\Container\ObjectSchema;
 use Blugen\Service\Lexicon\V1\Schema\Field\RefSchema;
 use Blugen\Service\Lexicon\V1\Schema\Field\TokenSchema;
 use Blugen\Service\Lexicon\V1\Schema\Field\UnionSchema;
